Some changes are implemented in Milestone 2

+++ TripNest_Backend/src/main/java/com/tripnest/backend/service/impl/ActivityServiceImpl.java
@@ -2,16 +2,24 @@ package com.tripnest.backend.service.impl;
 
 import java.util.List;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.tripnest.backend.common.ApiResponse;
 import com.tripnest.backend.dto.CreateActivityRequest;
+import com.tripnest.backend.dto.UpdateActivityRequest;
 import com.tripnest.backend.dto.response.ActivityResponse;
 import com.tripnest.backend.entity.Activity;
 import com.tripnest.backend.entity.Itinerary;
+import com.tripnest.backend.entity.Trip;
+import com.tripnest.backend.entity.User;
 import com.tripnest.backend.exception.ResourceNotFoundException;
 import com.tripnest.backend.repository.ActivityRepository;
 import com.tripnest.backend.repository.ItineraryRepository;
+import com.tripnest.backend.repository.TripRepository;
 import com.tripnest.backend.repository.UserRepository;
 import com.tripnest.backend.service.ActivityService;
 
@@ -24,38 +32,57 @@ public class ActivityServiceImpl implements ActivityService {
     private final ActivityRepository activityRepository;
     private final ItineraryRepository itineraryRepository;
     private final UserRepository userRepository;
-
+    private final TripRepository tripRepository;
+    
     @Override
+    @Transactional
     public ApiResponse<ActivityResponse> createActivity(
-            Long itineraryId,
+            Long tripId,
+            Integer dayNumber,
             CreateActivityRequest request) {
-    	Itinerary itinerary = itineraryRepository.findById(itineraryId)
-    	        .orElseThrow(() ->
-    	                new ResourceNotFoundException("Itinerary not found"));
 
-    	Activity activity = Activity.builder()
-    	        .title(request.getTitle())
-    	        .description(request.getDescription())
-    	        .activityTime(request.getActivityTime())
-    	        .activityType(request.getActivityType())
-    	        .itinerary(itinerary)
-    	        .build();
+        Trip trip = tripRepository.findById(tripId)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Trip not found"));
 
-    	activity = activityRepository.save(activity);
+        validateTripOwnership(trip);
 
-    	ActivityResponse response = ActivityResponse.builder()
-    	        .id(activity.getId())
-    	        .title(activity.getTitle())
-    	        .description(activity.getDescription())
-    	        .activityTime(activity.getActivityTime())
-    	        .activityType(activity.getActivityType())
-    	        .build();
+        Itinerary itinerary = itineraryRepository
+                .findByTripAndDayNumber(trip, dayNumber)
+                .orElseGet(() -> {
 
-    	return ApiResponse.<ActivityResponse>builder()
-    	        .success(true)
-    	        .message("Activity created successfully")
-    	        .data(response)
-    	        .build();
+                    Itinerary newItinerary = Itinerary.builder()
+                            .trip(trip)
+                            .dayNumber(dayNumber)
+                            .date(trip.getStartDate().plusDays(dayNumber - 1))
+                            .build();
+
+                    return itineraryRepository.save(newItinerary);
+                });
+
+        Activity activity = Activity.builder()
+                .title(request.getTitle())
+                .description(request.getDescription())
+                .activityTime(request.getActivityTime())
+                .activityType(request.getActivityType())
+                .itinerary(itinerary)
+                .build();
+
+        activity = activityRepository.save(activity);
+
+        ActivityResponse response = ActivityResponse.builder()
+                .id(activity.getId())
+                .title(activity.getTitle())
+                .description(activity.getDescription())
+                .activityTime(activity.getActivityTime())
+                .activityType(activity.getActivityType())
+                .build();
+
+        return ApiResponse.<ActivityResponse>builder()
+                .success(true)
+                .message("Activity created successfully")
+                .data(response)
+                .build();
     }
 
     @Override
@@ -65,6 +92,8 @@ public class ActivityServiceImpl implements ActivityService {
         Itinerary itinerary = itineraryRepository.findById(itineraryId)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Itinerary not found"));
+
+        validateTripOwnership(itinerary.getTrip());
 
         List<ActivityResponse> response = activityRepository
                 .findByItineraryOrderByActivityTimeAsc(itinerary)
@@ -88,18 +117,20 @@ public class ActivityServiceImpl implements ActivityService {
     @Override
     public ApiResponse<ActivityResponse> updateActivity(
             Long activityId,
-            CreateActivityRequest request) {
+            UpdateActivityRequest request) {
 
         Activity activity = activityRepository.findById(activityId)
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Activity not found"));
+
+        validateTripOwnership(activity.getItinerary().getTrip());
 
         activity.setTitle(request.getTitle());
         activity.setDescription(request.getDescription());
         activity.setActivityTime(request.getActivityTime());
         activity.setActivityType(request.getActivityType());
 
-        activityRepository.save(activity);
+        activity = activityRepository.save(activity);
 
         ActivityResponse response = ActivityResponse.builder()
                 .id(activity.getId())
@@ -116,6 +147,7 @@ public class ActivityServiceImpl implements ActivityService {
                 .build();
     }
 
+
     @Override
     public ApiResponse<String> deleteActivity(
             Long activityId) {
@@ -124,6 +156,8 @@ public class ActivityServiceImpl implements ActivityService {
                 .orElseThrow(() ->
                         new ResourceNotFoundException("Activity not found"));
 
+        validateTripOwnership(activity.getItinerary().getTrip());
+
         activityRepository.delete(activity);
 
         return ApiResponse.<String>builder()
@@ -131,5 +165,23 @@ public class ActivityServiceImpl implements ActivityService {
                 .message("Activity deleted successfully")
                 .data("Activity deleted successfully")
                 .build();
+    }
+
+    private User getCurrentUser() {
+        Authentication authentication =
+                SecurityContextHolder.getContext().getAuthentication();
+
+        String email = authentication.getName();
+
+        return userRepository.findByEmail(email)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User not found"));
+    }
+
+    private void validateTripOwnership(Trip trip) {
+        User currentUser = getCurrentUser();
+        if (!trip.getUser().getId().equals(currentUser.getId())) {
+            throw new AccessDeniedException("You do not have permission to modify this trip.");
+        }
     }
 }
